+++ src/main/java/io/github/astrapi69/swing/action/ReplaceContentAction.java
@@ -29,6 +29,7 @@ import java.awt.event.ActionEvent;
 import javax.swing.AbstractAction;
 import javax.swing.JComponent;
 
+import io.github.astrapi69.swing.component.replace.ReplaceContentExtensions;
 import lombok.NonNull;
 
 /**